+++ huaweicloud/provider.go
@@ -395,19 +395,19 @@ func Provider() *schema.Provider {
 			"huaweicloud_evs_volumes":        evs.DataSourceEvsVolumesV2(),
 			"huaweicloud_fgs_dependencies":   fgs.DataSourceFunctionGraphDependencies(),
 
-			"huaweicloud_gaussdb_cassandra_dedicated_resource": dataSourceGeminiDBDehResource(),
+			"huaweicloud_gaussdb_cassandra_dedicated_resource": gaussdb.DataSourceGeminiDBDehResource(),
 			"huaweicloud_gaussdb_cassandra_flavors":            gaussdb.DataSourceCassandraFlavors(),
 			"huaweicloud_gaussdb_nosql_flavors":                gaussdb.DataSourceGaussDBNoSQLFlavors(),
-			"huaweicloud_gaussdb_cassandra_instance":           dataSourceGeminiDBInstance(),
-			"huaweicloud_gaussdb_cassandra_instances":          dataSourceGeminiDBInstances(),
-			"huaweicloud_gaussdb_opengauss_instance":           dataSourceOpenGaussInstance(),
+			"huaweicloud_gaussdb_cassandra_instance":           gaussdb.DataSourceGeminiDBInstance(),
+			"huaweicloud_gaussdb_cassandra_instances":          gaussdb.DataSourceGeminiDBInstances(),
+			"huaweicloud_gaussdb_opengauss_instance":           gaussdb.DataSourceOpenGaussInstance(),
 			"huaweicloud_gaussdb_opengauss_instances":          gaussdb.DataSourceOpenGaussInstances(),
-			"huaweicloud_gaussdb_mysql_configuration":          dataSourceGaussdbMysqlConfigurations(),
-			"huaweicloud_gaussdb_mysql_dedicated_resource":     dataSourceGaussDBMysqlDehResource(),
-			"huaweicloud_gaussdb_mysql_flavors":                dataSourceGaussdbMysqlFlavors(),
-			"huaweicloud_gaussdb_mysql_instance":               dataSourceGaussDBMysqlInstance(),
-			"huaweicloud_gaussdb_mysql_instances":              dataSourceGaussDBMysqlInstances(),
-			"huaweicloud_gaussdb_redis_instance":               dataSourceGaussRedisInstance(),
+			"huaweicloud_gaussdb_mysql_configuration":          gaussdb.DataSourceGaussdbMysqlConfigurations(),
+			"huaweicloud_gaussdb_mysql_dedicated_resource":     gaussdb.DataSourceGaussDBMysqlDehResource(),
+			"huaweicloud_gaussdb_mysql_flavors":                gaussdb.DataSourceGaussdbMysqlFlavors(),
+			"huaweicloud_gaussdb_mysql_instance":               gaussdb.DataSourceGaussDBMysqlInstance(),
+			"huaweicloud_gaussdb_mysql_instances":              gaussdb.DataSourceGaussDBMysqlInstances(),
+			"huaweicloud_gaussdb_redis_instance":               gaussdb.DataSourceGaussRedisInstance(),
 
 			"huaweicloud_identity_role":        iam.DataSourceIdentityRoleV3(),
 			"huaweicloud_identity_custom_role": iam.DataSourceIdentityCustomRole(),
@@ -681,11 +681,11 @@ func Provider() *schema.Provider {
 			"huaweicloud_fgs_function":   fgs.ResourceFgsFunctionV2(),
 			"huaweicloud_fgs_trigger":    fgs.ResourceFunctionGraphTrigger(),
 
-			"huaweicloud_gaussdb_cassandra_instance": resourceGeminiDBInstanceV3(),
-			"huaweicloud_gaussdb_mysql_instance":     resourceGaussDBInstance(),
+			"huaweicloud_gaussdb_cassandra_instance": gaussdb.ResourceGeminiDBInstanceV3(),
+			"huaweicloud_gaussdb_mysql_instance":     gaussdb.ResourceGaussDBInstance(),
 			"huaweicloud_gaussdb_mysql_proxy":        gaussdb.ResourceGaussDBProxy(),
-			"huaweicloud_gaussdb_opengauss_instance": resourceOpenGaussInstance(),
-			"huaweicloud_gaussdb_redis_instance":     resourceGaussRedisInstanceV3(),
+			"huaweicloud_gaussdb_opengauss_instance": gaussdb.ResourceOpenGaussInstance(),
+			"huaweicloud_gaussdb_redis_instance":     gaussdb.ResourceGaussRedisInstanceV3(),
 
 			"huaweicloud_ges_graph": ResourceGesGraphV1(),
