+++ huaweicloud/services/acceptance/gaussdb/data_source_huaweicloud_gaussdb_mysql_instance_test.go
@@ -1,9 +1,10 @@
-package huaweicloud
+package gaussdb
 
 import (
 	"fmt"
 	"testing"
 
+	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/services/acceptance"
 	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/utils/fmtp"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/acctest"
@@ -15,8 +16,8 @@ func TestAccGaussdbMysqlInstanceDataSource_basic(t *testing.T) {
 	rName := fmt.Sprintf("tf-acc-test-%s", acctest.RandString(5))
 
 	resource.ParallelTest(t, resource.TestCase{
-		PreCheck:  func() { testAccPreCheck(t) },
-		Providers: testAccProviders,
+		PreCheck:          func() { acceptance.TestAccPreCheck(t) },
+		ProviderFactories: acceptance.TestAccProviderFactories,
 		Steps: []resource.TestStep{
 			{
 				Config: testAccGaussdbMysqlInstanceDataSource_basic(rName),
@@ -54,13 +55,13 @@ data "huaweicloud_networking_secgroup" "test" {
 }
 
 resource "huaweicloud_gaussdb_mysql_instance" "test" {
-  name        = "%s"
-  password    = "Test@123"
-  flavor      = "gaussdb.mysql.4xlarge.x86.4"
-  vpc_id      = huaweicloud_vpc.test.id
-  subnet_id   = huaweicloud_vpc_subnet.test.id
-
-  security_group_id = data.huaweicloud_networking_secgroup.test.id
+  name                  = "%s"
+  password              = "Test@12345678"
+  flavor                = "gaussdb.mysql.4xlarge.x86.4"
+  vpc_id                = huaweicloud_vpc.test.id
+  subnet_id             = huaweicloud_vpc_subnet.test.id
+  security_group_id     = data.huaweicloud_networking_secgroup.test.id
+  enterprise_project_id = "0"
 }
 
 data "huaweicloud_gaussdb_mysql_instance" "test" {
