+++ huaweicloud/services/acceptance/gaussdb/data_source_huaweicloud_gaussdb_mysql_instances_test.go
@@ -1,9 +1,10 @@
-package huaweicloud
+package gaussdb
 
 import (
 	"fmt"
 	"testing"
 
+	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/services/acceptance"
 	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/utils/fmtp"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/acctest"
@@ -15,8 +16,8 @@ func TestAccGaussdbMysqlInstancesDataSource_basic(t *testing.T) {
 	rName := fmt.Sprintf("tf-acc-test-%s", acctest.RandString(5))
 
 	resource.ParallelTest(t, resource.TestCase{
-		PreCheck:  func() { testAccPreCheck(t) },
-		Providers: testAccProviders,
+		PreCheck:          func() { acceptance.TestAccPreCheck(t) },
+		ProviderFactories: acceptance.TestAccProviderFactories,
 		Steps: []resource.TestStep{
 			{
 				Config: testAccGaussdbMysqlInstancesDataSource_basic(rName),
@@ -56,7 +57,7 @@ data "huaweicloud_networking_secgroup" "test" {
 
 resource "huaweicloud_gaussdb_mysql_instance" "test" {
   name                  = "%s"
-  password              = "Test@123"
+  password              = "Test@12345678"
   flavor                = "gaussdb.mysql.2xlarge.x86.4"
   vpc_id                = huaweicloud_vpc.test.id
   subnet_id             = huaweicloud_vpc_subnet.test.id
