+++ huaweicloud/services/gaussdb/data_source_huaweicloud_gaussdb_mysql_configuration.go
@@ -1,6 +1,7 @@
-package huaweicloud
+package gaussdb
 
 import (
+	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/common"
 	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/utils/fmtp"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
@@ -9,7 +10,7 @@ import (
 	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/config"
 )
 
-func dataSourceGaussdbMysqlConfigurations() *schema.Resource {
+func DataSourceGaussdbMysqlConfigurations() *schema.Resource {
 	return &schema.Resource{
 		Read: dataSourceGaussdbMysqlConfigurationsRead,
 
@@ -43,7 +44,7 @@ func dataSourceGaussdbMysqlConfigurations() *schema.Resource {
 func dataSourceGaussdbMysqlConfigurationsRead(d *schema.ResourceData, meta interface{}) error {
 	config := meta.(*config.Config)
 
-	client, err := config.GaussdbV3Client(GetRegion(d, config))
+	client, err := config.GaussdbV3Client(config.GetRegion(d))
 	if err != nil {
 		return fmtp.Errorf("Error creating HuaweiCloud GaussDB client: %s", err)
 	}
@@ -57,7 +58,7 @@ func dataSourceGaussdbMysqlConfigurationsRead(d *schema.ResourceData, meta inter
 			"Please change your search criteria and try again.")
 	}
 
-	if hasFilledOpt(d, "name") {
+	if common.HasFilledOpt(d, "name") {
 		var filteredConfigs []configurations.Configuration
 		for _, conf := range configsList {
 			if conf.Name == d.Get("name").(string) {
