+++ huaweicloud/services/gaussdb/data_source_huaweicloud_gaussdb_redis_instance.go
@@ -1,20 +1,24 @@
-package huaweicloud
+package gaussdb
 
 import (
+	"sort"
 	"strconv"
+	"strings"
 
 	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/utils/fmtp"
 	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/utils/logp"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 
-	"github.com/chnsz/golangsdk/openstack/taurusdb/v3/instances"
+	"github.com/chnsz/golangsdk/openstack/common/tags"
+	"github.com/chnsz/golangsdk/openstack/geminidb/v3/instances"
 	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/config"
+	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/utils"
 )
 
-func dataSourceGaussDBMysqlInstance() *schema.Resource {
+func DataSourceGaussRedisInstance() *schema.Resource {
 	return &schema.Resource{
-		Read: dataSourceGaussDBMysqlInstanceRead,
+		Read: dataSourceGaussRedisInstanceRead,
 
 		Schema: map[string]*schema.Schema{
 			"region": {
@@ -45,11 +49,11 @@ func dataSourceGaussDBMysqlInstance() *schema.Resource {
 				Type:     schema.TypeString,
 				Computed: true,
 			},
-			"security_group_id": {
+			"flavor": {
 				Type:     schema.TypeString,
 				Computed: true,
 			},
-			"configuration_id": {
+			"security_group_id": {
 				Type:     schema.TypeString,
 				Computed: true,
 			},
@@ -61,32 +65,40 @@ func dataSourceGaussDBMysqlInstance() *schema.Resource {
 				Type:     schema.TypeString,
 				Computed: true,
 			},
-			"time_zone": {
-				Type:     schema.TypeString,
-				Computed: true,
-			},
-			"availability_zone_mode": {
-				Type:     schema.TypeString,
-				Computed: true,
-			},
-			"master_availability_zone": {
-				Type:     schema.TypeString,
-				Computed: true,
-			},
 			"port": {
 				Type:     schema.TypeInt,
 				Computed: true,
 			},
-			"private_write_ip": {
+			"availability_zone": {
 				Type:     schema.TypeString,
 				Computed: true,
 			},
+			"node_num": {
+				Type:     schema.TypeInt,
+				Computed: true,
+			},
+			"volume_size": {
+				Type:     schema.TypeInt,
+				Computed: true,
+			},
+			"private_ips": {
+				Type:     schema.TypeList,
+				Computed: true,
+				Elem: &schema.Schema{
+					Type: schema.TypeString,
+				},
+			},
+
 			"datastore": {
 				Type:     schema.TypeList,
 				Computed: true,
 				Elem: &schema.Resource{
 					Schema: map[string]*schema.Schema{
 						"engine": {
+							Type:     schema.TypeString,
+							Computed: true,
+						},
+						"storage_engine": {
 							Type:     schema.TypeString,
 							Computed: true,
 						},
@@ -113,14 +125,6 @@ func dataSourceGaussDBMysqlInstance() *schema.Resource {
 					},
 				},
 			},
-			"read_replicas": {
-				Type:     schema.TypeInt,
-				Computed: true,
-			},
-			"flavor": {
-				Type:     schema.TypeString,
-				Computed: true,
-			},
 			"nodes": {
 				Type:     schema.TypeList,
 				Computed: true,
@@ -134,7 +138,7 @@ func dataSourceGaussDBMysqlInstance() *schema.Resource {
 							Type:     schema.TypeString,
 							Computed: true,
 						},
-						"type": {
+						"private_ip": {
 							Type:     schema.TypeString,
 							Computed: true,
 						},
@@ -142,8 +146,8 @@ func dataSourceGaussDBMysqlInstance() *schema.Resource {
 							Type:     schema.TypeString,
 							Computed: true,
 						},
-						"private_read_ip": {
-							Type:     schema.TypeString,
+						"support_reduce": {
+							Type:     schema.TypeBool,
 							Computed: true,
 						},
 						"availability_zone": {
@@ -153,19 +157,26 @@ func dataSourceGaussDBMysqlInstance() *schema.Resource {
 					},
 				},
 			},
+			"tags": {
+				Type:     schema.TypeMap,
+				Computed: true,
+				Elem: &schema.Schema{
+					Type: schema.TypeString,
+				},
+			},
 		},
 	}
 }
 
-func dataSourceGaussDBMysqlInstanceRead(d *schema.ResourceData, meta interface{}) error {
+func dataSourceGaussRedisInstanceRead(d *schema.ResourceData, meta interface{}) error {
 	config := meta.(*config.Config)
-	region := GetRegion(d, config)
-	client, err := config.GaussdbV3Client(region)
+	region := config.GetRegion(d)
+	client, err := config.GeminiDBV3Client(region)
 	if err != nil {
-		return fmtp.Errorf("Error creating HuaweiCloud GaussDB client: %s", err)
+		return fmtp.Errorf("Error creating HuaweiCloud GaussDB for Redis client: %s", err)
 	}
 
-	listOpts := instances.ListTaurusDBInstanceOpts{
+	listOpts := instances.ListGeminiDBInstanceOpts{
 		Name:     d.Get("name").(string),
 		VpcId:    d.Get("vpc_id").(string),
 		SubnetId: d.Get("subnet_id").(string),
@@ -176,7 +187,7 @@ func dataSourceGaussDBMysqlInstanceRead(d *schema.ResourceData, meta interface{}
 		return err
 	}
 
-	allInstances, err := instances.ExtractTaurusDBInstances(pages)
+	allInstances, err := instances.ExtractGeminiDBInstances(pages)
 	if err != nil {
 		return fmtp.Errorf("Unable to retrieve instances: %s", err)
 	}
@@ -191,90 +202,95 @@ func dataSourceGaussDBMysqlInstanceRead(d *schema.ResourceData, meta interface{}
 			" Please try a more specific search criteria")
 	}
 
-	instanceID := allInstances.Instances[0].Id
-	instance, err := instances.Get(client, instanceID).Extract()
-	if err != nil {
-		return err
-	}
+	instance := allInstances.Instances[0]
 
 	logp.Printf("[DEBUG] Retrieved Instance %s: %+v", instance.Id, instance)
 	d.SetId(instance.Id)
 
-	d.Set("region", region)
+	d.Set("region", instance.Region)
 	d.Set("name", instance.Name)
-	d.Set("status", instance.Status)
-	d.Set("mode", instance.Type)
 	d.Set("vpc_id", instance.VpcId)
 	d.Set("subnet_id", instance.SubnetId)
+	d.Set("status", instance.Status)
+	d.Set("mode", instance.Mode)
 	d.Set("security_group_id", instance.SecurityGroupId)
-	d.Set("configuration_id", instance.ConfigurationId)
 	d.Set("enterprise_project_id", instance.EnterpriseProjectId)
 	d.Set("db_user_name", instance.DbUserName)
-	d.Set("time_zone", instance.TimeZone)
-	d.Set("availability_zone_mode", instance.AZMode)
-	d.Set("master_availability_zone", instance.MasterAZ)
 
 	if dbPort, err := strconv.Atoi(instance.Port); err == nil {
 		d.Set("port", dbPort)
 	}
-	if len(instance.PrivateIps) > 0 {
-		d.Set("private_write_ip", instance.PrivateIps[0])
-	}
 
-	// set data store
-	dbList := make([]map[string]interface{}, 1)
+	dbList := make([]map[string]interface{}, 0, 1)
 	db := map[string]interface{}{
-		"version": instance.DataStore.Version,
+		"engine":         instance.DataStore.Type,
+		"version":        instance.DataStore.Version,
+		"storage_engine": instance.Engine,
 	}
-	// normalize engine
-	engine := instance.DataStore.Type
-	if engine == "GaussDB(for MySQL)" {
-		engine = "gaussdb-mysql"
-	}
-	db["engine"] = engine
-	dbList[0] = db
+	dbList = append(dbList, db)
 	d.Set("datastore", dbList)
 
-	// set nodes
-	flavor := ""
-	slave_count := 0
+	specCode := ""
+	wrongFlavor := "Inconsistent Flavor"
+	ipsList := []string{}
+	azList := []string{}
 	nodesList := make([]map[string]interface{}, 0, 1)
-	for _, raw := range instance.Nodes {
-		node := map[string]interface{}{
-			"id":                raw.Id,
-			"name":              raw.Name,
-			"status":            raw.Status,
-			"type":              raw.Type,
-			"availability_zone": raw.AvailabilityZone,
+	for _, group := range instance.Groups {
+		for _, Node := range group.Nodes {
+			node := map[string]interface{}{
+				"id":                Node.Id,
+				"name":              Node.Name,
+				"status":            Node.Status,
+				"private_ip":        Node.PrivateIp,
+				"support_reduce":    Node.SupportReduce,
+				"availability_zone": Node.AvailabilityZone,
+			}
+			if specCode == "" {
+				specCode = Node.SpecCode
+			} else if specCode != Node.SpecCode && specCode != wrongFlavor {
+				specCode = wrongFlavor
+			}
+			nodesList = append(nodesList, node)
+			azList = append(azList, Node.AvailabilityZone)
+			// Only return Node private ips which doesn't support reduce
+			if !Node.SupportReduce {
+				ipsList = append(ipsList, Node.PrivateIp)
+			}
 		}
-		if len(raw.PrivateIps) > 0 {
-			node["private_read_ip"] = raw.PrivateIps[0]
+		if volSize, err := strconv.Atoi(group.Volume.Size); err == nil {
+			d.Set("volume_size", volSize)
 		}
-		nodesList = append(nodesList, node)
-		if raw.Type == "slave" && raw.Status == "ACTIVE" {
-			slave_count += 1
-		}
-		if flavor == "" {
-			flavor = raw.Flavor
+		if specCode != "" {
+			logp.Printf("[DEBUG] Node SpecCode: %s", specCode)
+			d.Set("flavor", specCode)
 		}
 	}
 	d.Set("nodes", nodesList)
-	d.Set("read_replicas", slave_count)
-	if flavor != "" {
-		logp.Printf("[DEBUG] Node Flavor: %s", flavor)
-		d.Set("flavor", flavor)
-	}
+	d.Set("private_ips", ipsList)
 
-	// set backup_strategy
-	backupStrategyList := make([]map[string]interface{}, 1)
+	//remove duplicate az
+	azList = utils.RemoveDuplicateElem(azList)
+	sort.Strings(azList)
+	d.Set("availability_zone", strings.Join(azList, ","))
+	d.Set("node_num", len(nodesList))
+
+	backupStrategyList := make([]map[string]interface{}, 0, 1)
 	backupStrategy := map[string]interface{}{
 		"start_time": instance.BackupStrategy.StartTime,
+		"keep_days":  instance.BackupStrategy.KeepDays,
 	}
-	if days, err := strconv.Atoi(instance.BackupStrategy.KeepDays); err == nil {
-		backupStrategy["keep_days"] = days
-	}
-	backupStrategyList[0] = backupStrategy
+	backupStrategyList = append(backupStrategyList, backupStrategy)
 	d.Set("backup_strategy", backupStrategyList)
+
+	//save geminidb tags
+	if resourceTags, err := tags.Get(client, "instances", d.Id()).Extract(); err == nil {
+		tagmap := utils.TagsToMap(resourceTags.Tags)
+		if err := d.Set("tags", tagmap); err != nil {
+			return fmtp.Errorf("Error saving tags to state for geminidb (%s): %s", d.Id(), err)
+		}
+	} else {
+		logp.Printf("[WARN] Error fetching tags of geminidb (%s): %s", d.Id(), err)
+	}
 
 	return nil
 }
