+++ huaweicloud/services/gaussdb/resource_huaweicloud_gaussdb_mysql_instance.go
@@ -1,4 +1,4 @@
-package huaweicloud
+package gaussdb
 
 import (
 	"context"
@@ -24,7 +24,7 @@ import (
 	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/utils/logp"
 )
 
-func resourceGaussDBInstance() *schema.Resource {
+func ResourceGaussDBInstance() *schema.Resource {
 	return &schema.Resource{
 		Create: resourceGaussDBInstanceCreate,
 		Update: resourceGaussDBInstanceUpdate,
@@ -208,7 +208,7 @@ func resourceGaussDBInstance() *schema.Resource {
 				Optional: true,
 			},
 			// only supported in some regions, so it's not shown in the doc
-			"tags": tagsSchema(),
+			"tags": common.TagsSchema(),
 
 			"proxy_address": {
 				Type:     schema.TypeString,
@@ -351,13 +351,13 @@ func GaussDBInstanceStateRefreshFunc(client *golangsdk.ServiceClient, instanceID
 
 func resourceGaussDBInstanceCreate(d *schema.ResourceData, meta interface{}) error {
 	config := meta.(*config.Config)
-	client, err := config.GaussdbV3Client(GetRegion(d, config))
+	client, err := config.GaussdbV3Client(config.GetRegion(d))
 	if err != nil {
 		return fmtp.Errorf("error creating HuaweiCloud GaussDB client: %s ", err)
 	}
 
 	// If force_import set, try to import it instead of creating
-	if hasFilledOpt(d, "force_import") {
+	if common.HasFilledOpt(d, "force_import") {
 		logp.Printf("[DEBUG] Gaussdb mysql instance force_import is set, try to import it instead of creating")
 		listOpts := instances.ListTaurusDBInstanceOpts{
 			Name: d.Get("name").(string),
@@ -382,12 +382,12 @@ func resourceGaussDBInstanceCreate(d *schema.ResourceData, meta interface{}) err
 	createOpts := instances.CreateTaurusDBOpts{
 		Name:                d.Get("name").(string),
 		Flavor:              d.Get("flavor").(string),
-		Region:              GetRegion(d, config),
+		Region:              config.GetRegion(d),
 		VpcId:               d.Get("vpc_id").(string),
 		SubnetId:            d.Get("subnet_id").(string),
 		SecurityGroupId:     d.Get("security_group_id").(string),
 		ConfigurationId:     d.Get("configuration_id").(string),
-		EnterpriseProjectId: GetEnterpriseProjectID(d, config),
+		EnterpriseProjectId: config.GetEnterpriseProjectID(d),
 		DedicatedResourceId: d.Get("dedicated_resource_id").(string),
 		TimeZone:            d.Get("time_zone").(string),
 		SlaveCount:          d.Get("read_replicas").(int),
@@ -410,7 +410,7 @@ func resourceGaussDBInstanceCreate(d *schema.ResourceData, meta interface{}) err
 		createOpts.MasterAZ = v.(string)
 	}
 
-	if hasFilledOpt(d, "volume_size") {
+	if common.HasFilledOpt(d, "volume_size") {
 		volume := &instances.VolumeOpt{
 			Size: d.Get("volume_size").(int),
 		}
@@ -460,7 +460,7 @@ func resourceGaussDBInstanceCreate(d *schema.ResourceData, meta interface{}) err
 
 	// PrePaid
 	if d.Get("charging_mode") == "prePaid" {
-		if err := validatePrePaidChargeInfo(d); err != nil {
+		if err := common.ValidatePrePaidChargeInfo(d); err != nil {
 			return err
 		}
 
@@ -532,7 +532,7 @@ func resourceGaussDBInstanceCreate(d *schema.ResourceData, meta interface{}) err
 		}
 	}
 
-	if hasFilledOpt(d, "backup_strategy") {
+	if common.HasFilledOpt(d, "backup_strategy") {
 		var updateOpts backups.UpdateOpts
 		backupRaw := d.Get("backup_strategy").([]interface{})
 		rawMap := backupRaw[0].(map[string]interface{})
@@ -549,7 +549,7 @@ func resourceGaussDBInstanceCreate(d *schema.ResourceData, meta interface{}) err
 		}
 	}
 
-	if hasFilledOpt(d, "proxy_flavor") {
+	if common.HasFilledOpt(d, "proxy_flavor") {
 		proxyOpts := instances.ProxyOpts{
 			Flavor:  d.Get("proxy_flavor").(string),
 			NodeNum: d.Get("proxy_node_num").(int),
@@ -580,7 +580,7 @@ func resourceGaussDBInstanceCreate(d *schema.ResourceData, meta interface{}) err
 
 func resourceGaussDBInstanceRead(d *schema.ResourceData, meta interface{}) error {
 	config := meta.(*config.Config)
-	region := GetRegion(d, config)
+	region := config.GetRegion(d)
 	client, err := config.GaussdbV3Client(region)
 	if err != nil {
 		return fmtp.Errorf("error creating HuaweiCloud GaussDB client: %s", err)
@@ -589,7 +589,7 @@ func resourceGaussDBInstanceRead(d *schema.ResourceData, meta interface{}) error
 	instanceID := d.Id()
 	instance, err := instances.Get(client, instanceID).Extract()
 	if err != nil {
-		return CheckDeleted(d, err, "GaussDB instance")
+		return common.CheckDeleted(d, err, "GaussDB instance")
 	}
 	if instance.Id == "" {
 		d.SetId("")
@@ -750,11 +750,11 @@ func resourceGaussDBInstanceRead(d *schema.ResourceData, meta interface{}) error
 
 func resourceGaussDBInstanceUpdate(d *schema.ResourceData, meta interface{}) error {
 	config := meta.(*config.Config)
-	client, err := config.GaussdbV3Client(GetRegion(d, config))
+	client, err := config.GaussdbV3Client(config.GetRegion(d))
 	if err != nil {
 		return fmtp.Errorf("error creating HuaweiCloud GaussDB client: %s ", err)
 	}
-	bssClient, err := config.BssV2Client(GetRegion(d, config))
+	bssClient, err := config.BssV2Client(config.GetRegion(d))
 	if err != nil {
 		return fmtp.Errorf("error creating HuaweiCloud bss V2 client: %s", err)
 	}
@@ -975,7 +975,7 @@ func resourceGaussDBInstanceUpdate(d *schema.ResourceData, meta interface{}) err
 	}
 
 	if d.HasChange("proxy_flavor") {
-		if hasFilledOpt(d, "proxy_flavor") {
+		if common.HasFilledOpt(d, "proxy_flavor") {
 			proxyOpts := instances.ProxyOpts{
 				Flavor:  d.Get("proxy_flavor").(string),
 				NodeNum: d.Get("proxy_node_num").(int),
@@ -1004,7 +1004,7 @@ func resourceGaussDBInstanceUpdate(d *schema.ResourceData, meta interface{}) err
 
 	if d.HasChange("proxy_node_num") {
 		oldnum, newnum := d.GetChange("proxy_node_num")
-		if oldnum.(int) != 0 && newnum.(int) > oldnum.(int) && hasFilledOpt(d, "proxy_flavor") {
+		if oldnum.(int) != 0 && newnum.(int) > oldnum.(int) && common.HasFilledOpt(d, "proxy_flavor") {
 			enlarge_size := newnum.(int) - oldnum.(int)
 			enlargeProxyOpts := instances.EnlargeProxyOpts{
 				NodeNum: enlarge_size,
@@ -1045,24 +1045,24 @@ func resourceGaussDBInstanceUpdate(d *schema.ResourceData, meta interface{}) err
 
 func resourceGaussDBInstanceDelete(d *schema.ResourceData, meta interface{}) error {
 	config := meta.(*config.Config)
-	client, err := config.GaussdbV3Client(GetRegion(d, config))
+	client, err := config.GaussdbV3Client(config.GetRegion(d))
 	if err != nil {
 		return fmtp.Errorf("error creating HuaweiCloud GaussDB client: %s ", err)
 	}
 
 	instanceId := d.Id()
 	if d.Get("charging_mode") == "prePaid" {
-		if err := UnsubscribePrePaidResource(d, config, []string{instanceId}); err != nil {
+		if err := common.UnsubscribePrePaidResource(d, config, []string{instanceId}); err != nil {
 			// try to delete the instance directly if unsubscribing failed
 			res := instances.Delete(client, instanceId)
 			if res.Err != nil {
-				return CheckDeleted(d, res.Err, "GaussDB instance")
+				return common.CheckDeleted(d, res.Err, "GaussDB instance")
 			}
 		}
 	} else {
 		result := instances.Delete(client, instanceId)
 		if result.Err != nil {
-			return CheckDeleted(d, result.Err, "GaussDB instance")
+			return common.CheckDeleted(d, result.Err, "GaussDB instance")
 		}
 	}
